feat: add project detail, milestones, outcomes, and notes pages with data fetching and layout
