+++ app/[language]/(protected)/(default-layout)/projects/notes/page.tsx
